added host config to automatically use origin as hostname. For development, auto needs to be disabled renamed host config file
