+++ src/components/Explore.jsx
@@ -12,7 +12,7 @@ import makeStyles from '@mui/styles/makeStyles';
 
 // utility
 import {api, endpoints} from '../api';
-import {host} from '../constants';
+import {host} from '../host';
 import {isEmpty} from 'lodash';
 import FilterCard from './FilterCard.jsx';
 //
@@ -43,7 +43,7 @@ function Explore ({ keywords, data, updateData }) {
     const classes = useStyles();
 
     useEffect(() => {
-        api.get(`${host}${endpoints.imageList}`)
+        api.get(`${host()}${endpoints.imageList}`)
           .then(response => {
             if (response.data && response.data.data) {
                 let imageList = response.data.data.ImageListWithLatestTag;
