--- src/constants.js
@@ -1,3 +1,0 @@
-const host = 'http://localhost:5000'
-
-export {host};
